remove compilation.config
